moved some code into QtInfo, and added a line to .gitignore
